+++ packages/retag/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @definitelytyped/retag
 
+## 0.0.200
+
+### Patch Changes
+
+- @definitelytyped/definitions-parser@0.0.200
+
+## 0.0.199
+
+### Patch Changes
+
+- @definitelytyped/definitions-parser@0.0.199
+
+## 0.0.198
+
+### Patch Changes
+
+- Updated dependencies [4522dfba]
+  - @definitelytyped/typescript-versions@0.0.184
+  - @definitelytyped/definitions-parser@0.0.198
+  - @definitelytyped/utils@0.0.191
+
+## 0.0.197
+
+### Patch Changes
+
+- b980f717: Rename branch to main
+- Updated dependencies [b980f717]
+  - @definitelytyped/typescript-versions@0.0.183
+  - @definitelytyped/definitions-parser@0.0.197
+  - @definitelytyped/utils@0.0.190
+
 ## 0.0.196
 
 ### Patch Changes