+++ packages/typescript-versions/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @definitelytyped/typescript-versions
 
+## 0.0.184
+
+### Patch Changes
+
+- 4522dfba: Improve types of versions
+
+## 0.0.183
+
+### Patch Changes
+
+- b980f717: Rename branch to main
+
 ## 0.0.182
 
 ### Patch Changes